+++ SourceryRuntime/Sources/TemplateContext.swift
@@ -8,7 +8,7 @@ import Foundation
 /// :nodoc:
 @objcMembers public final class TemplateContext: NSObject, SourceryModel {
     public let types: Types
-    public let arguments: [String: NSObject]
+    public let argument: [String: NSObject]
 
     // sourcery: skipDescription
     public var type: [String: Type] {
@@ -17,20 +17,20 @@ import Foundation
 
     public init(types: Types, arguments: [String: NSObject]) {
         self.types = types
-        self.arguments = arguments
+        self.argument = arguments
     }
 
     // sourcery:inline:TemplateContext.AutoCoding
         /// :nodoc:
         required public init?(coder aDecoder: NSCoder) {
             guard let types: Types = aDecoder.decode(forKey: "types") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["types"])); fatalError() }; self.types = types
-            guard let arguments: [String: NSObject] = aDecoder.decode(forKey: "arguments") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["arguments"])); fatalError() }; self.arguments = arguments
+            guard let argument: [String: NSObject] = aDecoder.decode(forKey: "argument") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["argument"])); fatalError() }; self.argument = argument
         }
 
         /// :nodoc:
         public func encode(with aCoder: NSCoder) {
             aCoder.encode(self.types, forKey: "types")
-            aCoder.encode(self.arguments, forKey: "arguments")
+            aCoder.encode(self.argument, forKey: "argument")
         }
     // sourcery:end
 
@@ -38,7 +38,7 @@ import Foundation
         return [
             "types": types,
             "type": types.typesByName,
-            "argument": arguments
+            "argument": argument
         ]
     }
 
@@ -57,7 +57,7 @@ import Foundation
                 "implementing": types.implementing
             ],
             "type": types.typesByName,
-            "argument": arguments
+            "argument": argument
         ]
     }